Add ValheimVikiApp main component, Refactored SharedTransitionLayout (It work but still WIP)
