+++ app/.idea/workspace.xml
@@ -1,0 +1,77 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="2af9f4c2-438b-480f-a4e1-63e6586e97ff" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/navigation/ChildNavGraph.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/navigation/ChildNavGraph.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/navigation/NavGraph.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/navigation/NavGraph.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/detail/biome/BiomeDetailScreen.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/detail/biome/BiomeDetailScreen.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/detail/creature/CreatureDetailScreen.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/detail/creature/CreatureDetailScreen.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/home/HomeContent.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/home/HomeContent.kt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/home/HomeScreen.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/rabbitv/valheimviki/presentation/home/HomeScreen.kt" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$/.." />
+  </component>
+  <component name="GitHubPullRequestSearchHistory">{
+  &quot;lastFilter&quot;: {
+    &quot;state&quot;: &quot;OPEN&quot;,
+    &quot;assignee&quot;: &quot;DwMichael&quot;
+  }
+}</component>
+  <component name="GithubPullRequestsUISettings">{
+  &quot;selectedUrlAndAccountId&quot;: {
+    &quot;url&quot;: &quot;https://github.com/DwMichael/ValheimVikiApp.git&quot;,
+    &quot;accountId&quot;: &quot;121e17dc-cac1-4c84-8cd0-6b7225081e65&quot;
+  }
+}</component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 5
+}</component>
+  <component name="ProjectId" id="2vMyywLrCZdIQFpycT7k2vTKppE" />
+  <component name="ProjectViewState">
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.git.unshallow&quot;: &quot;true&quot;,
+    &quot;git-widget-placeholder&quot;: &quot;feature/BiomeDetailScreen&quot;,
+    &quot;kotlin-language-version-configured&quot;: &quot;true&quot;,
+    &quot;nodejs_package_manager_path&quot;: &quot;npm&quot;,
+    &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
+  }
+}</component>
+  <component name="SharedIndexes">
+    <attachedChunks>
+      <set>
+        <option value="bundled-jdk-9823dce3aa75-a94e463ab2e7-intellij.indexing.shared.core-IU-243.26053.27" />
+        <option value="bundled-js-predefined-d6986cc7102b-1632447f56bf-JavaScript-IU-243.26053.27" />
+      </set>
+    </attachedChunks>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="2af9f4c2-438b-480f-a4e1-63e6586e97ff" name="Changes" comment="" />
+      <created>1743969496240</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1743969496240</updated>
+      <workItem from="1743969497755" duration="9000" />
+      <workItem from="1744011333192" duration="134000" />
+      <workItem from="1744014361633" duration="9000" />
+    </task>
+    <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+</project>
