+++ app/src/main/java/com/rabbitv/valheimviki/MainActivity.kt
@@ -7,9 +7,6 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
-import com.rabbitv.valheimviki.navigation.SetupNavGraph
 import com.rabbitv.valheimviki.ui.theme.ValheimVikiAppTheme
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -17,17 +14,12 @@ import dagger.hilt.android.AndroidEntryPoint
 @AndroidEntryPoint
 class MainActivity() : ComponentActivity() {
 
-    private lateinit var navController: NavHostController
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         installSplashScreen()
         enableEdgeToEdge()
         setContent {
-            ValheimVikiAppTheme {
-                navController = rememberNavController()
-                SetupNavGraph(navController = navController)
-            }
+            ValheimVikiApp()
         }
     }
 }
